requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.6
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.2
lxml==4.6.3
pkg_resources==0.0.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
pytz==2021.1
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.7
